requirements.txt
streamlit
langchain
langchain-community
chromadb
tiktoken
# pysqlite3-binary
pypdf
langchain-openai
sentence-transformers
openai 
python-dotenv
protobuf==3.20.3